Add mlton-default.basis
Add a copy of MLton's default environment (implicitly used when
compiling a `.sml` file); it is defined as

    $(SML_LIB)/basis/unsafe.mlb
    $(SML_LIB)/basis/sml-nj.mlb
    $(SML_LIB)/basis/mlton.mlb
    $(SML_LIB)/basis/basis.mlb

That is, it extends the Standard ML Basis Library with
`structure Unsafe: UNSAFE`, `structure SMLofNJ: SML_OF_NJ`, and
`structure MLton: MLTON` (plus supporting signatures).

Could be used as the default basis when a better choice is not
available.